pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noleak"
version = "1.0.0"
description = "A DevSecOps tool for scanning hardcoded secrets in source code"
authors = [{name = "NoLeak Team", email = "security@noleak.dev"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "PyYAML>=6.0"
]

[project.urls]
Homepage = "https://github.com/Seichs/NoLeak"
"Bug Reports" = "https://github.com/Seichs/NoLeak/issues"
"Source" = "https://github.com/Seichs/NoLeak"

[project.scripts]
noleak = "noleak.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=noleak --cov-report=term-missing"

[tool.coverage.run]
source = ["noleak"]
omit = ["*/tests/*", "setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
